==> sample_reproduction/01_generate_hypothetical_terms.ipynb <==
requirements:

* llm access

outputs:

* topics.json
* term_candidates.json
* hypothetical_term_candidates.csv

# import external libraries
import sys
import os
import json
import pandas as pd
import numpy as np
from io import StringIO

# import local modules
current_dir = os.path.dirname(os.path.abspath('__file__'))
parent_dir = os.path.dirname(current_dir)
data_dir = os.path.join(parent_dir, 'data')

sys.path.append(os.path.join(parent_dir))

from src.searchUnexistent import Searcher
from src.utilities import get_strtime
from prompts import templates
import settings

### Topic Generation Prompt

topic_generator = "What are the most popular 20 topics on the internet?"

### Term Generation Prompt

def generate_term_generation_prompt(topics:list):
    for topic in topics:
        yield templates.term_generator.format(topic=topic)

topics = json.load(open(os.path.join(data_dir, 'intermediate', 'topics.json')))
prompt_gen = generate_term_generation_prompt(topics)

print(next(prompt_gen))

### Transforming Term Candidates to Dataframe

term_candidates = json.load(open(os.path.join(data_dir, 'intermediate', 'term_candidates.json')))

data = []
for index, item in enumerate(term_candidates):
    terms = item.split("/n/n/n")
    for term in terms:
        term_title, definition = term.split(":")
        data.append({"term": term_title, "explanation": definition, "topic": topics[index]})

term_candidates_df = pd.DataFrame(data)
term_candidates_df["exists"]= np.nan

term_candidates_df

### Verifying Hypothetical Terms

searcher = Searcher(settings)

def search_nonexistent(df: pd.DataFrame):
    for index, row in df.iterrows():
        if not np.isnan(row["exists"]):
            continue
        does_exists = searcher.exists()
        df.loc[index, 'exists'] = does_exists

# testing searcher class
searcher.search("Nano-Sync Fusion Technology")

term_candidates_path = os.path.join(data_dir, "intermediate", "hypothetical_term_candidates.csv")
term_candidates_df = pd.read_csv(term_candidates_path)

search_nonexistent(term_candidates_df)

term_candidates_path = os.path.join(data_dir, 'intermediate', f"term_candidates_{get_strtime()}.csv")
term_candidates_df.to_csv(term_candidates_path, index=False)
