==> sample_reproduction/04_vectorize_wiki_entries.ipynb <==
Requirements:
* Wikipedia on NoSql DB (Mongo DB)
* Vector Database (Milvus)

Outputs:
* Wikipedia titles vector database
* Wikipedia definitions vector database

# import external libraries
import os
import sys

# import local modules
current_dir = os.path.dirname(os.path.abspath('__file__'))
parent_dir = os.path.dirname(current_dir)
data_dir = os.path.join(parent_dir, 'data')

sys.path.append(os.path.join(parent_dir))

import settings
from wiki.embeddings import EmbeddingTextIterator, EmbeddingTitleIterator

#### Vectorize Titles

embeddingTitleIterator = EmbeddingTitleIterator(settings)

embeddingTitleIterator.insert()

#### Vectorize Text

embeddingTextIterator = EmbeddingTextIterator(settings)

embeddingTextIterator.insert()
